refactor: extract login logic into LoginContainer component for improved code organization and maintainability
