+++ src/hooks/useAuth.ts
@@ -1,12 +1,17 @@
 import { useSelector, useDispatch } from 'react-redux';
 import { RootState, AppDispatch } from '../lib/store';
-import { login, logout } from '../lib/store/authSlice';
-import { useCallback } from 'react';
-import { LoginCredentials } from '@/lib/store/types';
+import { logout } from '@/lib/store/authSlice';
 
 export const useAuth = () => {
   const dispatch = useDispatch<AppDispatch>();
   const auth = useSelector((state: RootState) => state.auth);
 
-  
+  const logoutUser = () => {
+    dispatch(logout());
+  };
+
+  return {
+    ...auth,
+    logoutUser,
+  };
 };
